+++ src/test/java/transformxml.java
@@ -1,8 +1,10 @@
-package test;
+
 
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Writer;
+
+import org.jaxen.dom.DocumentNavigator;
 
 import be.baur.sda.DataNode;
 import be.baur.sda.SDA;
@@ -10,15 +12,16 @@ import be.baur.sdt.SDT;
 import be.baur.sdt.TransformContext;
 import be.baur.sdt.transform.Transform;
 
-public final class TestTransform {
+public final class transformxml {
 	
 	public static void main(String[] args) throws Exception {
 		
-		InputStream in = TestTransform.class.getResourceAsStream("/example.sdt");
+		InputStream in = transformxml.class.getResourceAsStream("/example-xml.sdt");
 		Transform tran = SDT.parse(new InputStreamReader(in, "UTF-8"));
 		
-		String file = TestTransform.class.getResource("/example.sda").getFile();
+		String file = transformxml.class.getResource("/example.xml").getFile();
 		TransformContext c = new TransformContext.Builder() //.setWriter(SDT.nullWriter())
+			.setNavigator(DocumentNavigator.getInstance())
 			.setStringParameter("filename", file).build();
 		Writer w = c.getWriter();